main_menu working
