+++ Menu/menu.py
@@ -1,70 +1,74 @@
 from rich.console import Console
 from rich.panel import Panel
 from rich.align import Align
+from rich.text import Text
+from rich import box
 from pynput import keyboard
 import os
+import sys
+import time
 
-#Esto nos sirve para llamar a las distitas funciones que se encuentran en estas carpetas
-def enlace_opciones():
-    sys.path.append("Nuevo juego")
-    sys.path.append("Cargar partida")
-    sys.path.append("Salir")
-    try:
-        import characters as cha
-        import saves as sav
-    except ModuleNotFoundError:
-        print("Couldn't find the necessary functions, closing the game...")
-        time.sleep(2)
-        sys.exit()
 
 console = Console()
-game_window = None
 
 class Menu():
-    def __init__(self, text, options, console, game_window):
-        self.text = text
-        self.options = options
-        self.current_option = 0
-        self.choice_made = False
-        self.console = console
-        self.game_window = game_window
+    def __init__(self, text, options, console):
+        self.text = text                        #Titulo
+        self.options = options                  #Opciones
+        self.current_option = 0                 #Indice de las opciones
+        self.choice_made = False                #Si se selecciona una opcion, por ende arranca en falso
+        self.console = console                  #Nos permite manipular la terminal      
+
  
-        #Printea El titulo
+    #Printea El titulo y lo posiciona a nuestro gusto
     def print_menu(self):
-        self.console.print(Panel(self.text,width=64),justify="center")
-        #Printea Las opciones
+        self.console.print(Panel(self.text, width=64), justify="center")
+
+    #Printea Las opciones y las pocisiona a nuestro gusto
     def print_option(self):
-        current_option_key = list(self.options.keys())[self.current_option]
+        current_option_key = self.current_option #Convierte el Diccionario en Lista
         options_text = ""
-        for option in self.options:
-            if option == current_option_key:
-                options_text = options_text + "> " + self.options[option] + " <\n"
-            else:
-                options_text = options_text + self.options[option] + "\n"
-        options_text = options_text.rstrip("\n")
+        for option in self.options: #Si coincide el indice de la opcion
+           if option == current_option_key:
+              options_text += f">[bold red on white] {self.options[option]} [/bold red on white]<\n"       #Muestra la opcion seleccionada
+           else:
+                options_text += self.options[option] + "\n"     #Printea las otras opciones no seleccionadas 
+        options_text = options_text.rstrip("\n") #Saca el ultimo salto de linea 
 
-        self.console.print(
-            Panel(Text(options_text, justify="center"), box=box.SIMPLE, width=64),
-            justify="center",
-        )
+        self.console.print(Panel(Text.from_markup(options_text, justify="center"), box=box.SIMPLE, width=64),justify="center",)
+        #Modifica la ubicacion de las opciones en la terminal
 
-        #Texto del Juego.El titulo-una breve descripcion y las opciones a elegir
-    def text_menu(self):
-        text = (f"[bold red] {"Divine Light"} [/bold red]\n [yellow]{"This Game is Awesome"} [/yellow]")           #Titulo con una breve descripcion
+    #Mostrar el menú completo
+    def show(self):
+        os.system("cls")
+        self.print_menu()
+        self.print_option()
 
-        options = { 0 : "Nuevo Juego",
-                    1 : "Cargar Partida",    
-                    2 : "Salir del Juego"}   #Opciones
+    def controll_keyboard(self,tecla):
+            indice = self.current_option
 
-        os.system("clear")
+            if tecla == keyboard.Key.up:
+                    self.current_option = (self.current_option - 1) % len(self.options)
+                    self.show()
+            elif tecla == keyboard.Key.down:
+                    self.current_option = (self.current_option + 1) % len(self.options)
+                    self.show()
+            elif tecla == keyboard.Key.enter:
+                    self.choice_made = True
+                    os.system("cls")
+                    console.print(f"[bold green]Seleccionaste: {self.options[self.current_option]} [/bold green]")
 
-        #Crear el Menu
-        starting_menu = Menu(text,options,console,game_window)
+#Texto del Juego.El titulo-una breve descripcion y las opciones a elegir
+text = "[bold red] Divine Light [/bold red]\n[yellow] This Game is Awesome [/yellow]"
 
-        #Mostrar el menu
-        os.system("clear")
-        starting_menu.print_menu()
-        starting_menu.print_option()
+options = { 
+    0 : "Nuevo Juego",
+    1 : "Cargar Partida",    
+    2 : "Salir del Juego"}
 
-Menu_Principal = Menu(text,options,console,game_window)
-Menu_Principal.text_menu()
+#Crear el Menu Principal
+Menu_Principal = Menu(text, options, console)
+Menu_Principal.show()
+
+with keyboard.Listener(on_press=Menu_Principal.controll_keyboard) as listener:
+    listener.join()
